% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wape.data.frame.R
\name{wape.data.frame}
\alias{wape.data.frame}
\title{WAPE data.frame}
\usage{
\method{wape}{data.frame}(data, truth, estimate, na_rm = TRUE, case_weights = NULL, ...)
}
\arguments{
\item{data}{\itemize{
\item data.frame
}}

\item{truth}{\itemize{
\item name variable
}}

\item{estimate}{\itemize{
\item predicted
}}

\item{na_rm}{\itemize{
\item Removing missing values
}}

\item{case_weights}{\itemize{
\item weights
}}

\item{...}{\itemize{
\item Extra options
}}
}
\value{
\itemize{
\item metric value
}
}
\description{
WAPE data.frame
}
