% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wape.R
\name{wape}
\alias{wape}
\title{Method wape}
\usage{
wape(data, ...)
}
\arguments{
\item{data}{\itemize{
\item data.frame
}}

\item{...}{\itemize{
\item Extra options
}}
}
\value{
\itemize{
\item method
}
}
\description{
Method wape
}
