% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wape_impl.R
\name{wape_impl}
\alias{wape_impl}
\title{Create wape_impl}
\usage{
wape_impl(truth, estimate, case_weights = NULL)
}
\arguments{
\item{truth}{\itemize{
\item vector
}}

\item{estimate}{\itemize{
\item vector predicted
}}

\item{case_weights}{\itemize{
\item vector
}}
}
\value{
\itemize{
\item 
\itemize{
\item vector
}
}
}
\description{
Create wape_impl
}
