% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wape_vec.R
\name{wape_vec}
\alias{wape_vec}
\title{WAPE vector}
\usage{
wape_vec(truth, estimate, na_rm = TRUE, case_weights = NULL, ...)
}
\arguments{
\item{truth}{\itemize{
\item Vector
}}

\item{estimate}{\itemize{
\item Vector predicted
}}

\item{na_rm}{\itemize{
\item Removing missing values
}}

\item{case_weights}{\itemize{
\item Vector
}}

\item{...}{\itemize{
\item Extra options
}}
}
\value{
\itemize{
\item metric value
}
}
\description{
WAPE vector
}
